backend/requirements.txt
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.9.2
asn1crypto==1.5.1
billiard==4.2.2
celery==5.5.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
dj-database-url==2.1.0
dj3-cloudinary-storage==0.0.6
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==25.2
django-storages==1.14.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
flower==2.0.1
gunicorn==21.2.0
html5lib==1.1
humanize==4.13.0
idna==3.7
inflection==0.5.1
kombu==5.5.4
lxml==5.2.1
oscrypto==1.3.0
packaging==25.0
pdf2image==1.16.3
Pillow==10.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pyHanko==0.24.0
pyhanko-certvalidator==0.26.3
pyjokes==0.8.3
PyJWT==2.8.0
pypdf==4.2.0
pypng==0.20220715.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==6.4.0
reportlab==4.0.9
requests==2.31.0
resend==2.1.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.5.2
typing_extensions==4.11.0
tzdata==2025.2
tzlocal==5.2
uritemplate==4.2.0
uritools==4.0.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
yapf==0.33.0
